Fixed bug where FilterRequest has Set of Integer instead of Set of Long

+++ src/main/java/de/yalama/hackerrankindexer/Submission/Model/FilterRequest.java
@@ -18,5 +18,5 @@ public class FilterRequest {
      * 3 is most recent passed
      */
     private int mode;
-    private Set<Integer> languageIDs;
+    private Set<Long> languageIDs;
 }
